# src/controllers/authController.js
import {
  registerUser,
  verifyUserEmail,
  resendVerificationUserEmail,
  loginUser,
  refreshSession,
} from "../services/auth.js";
import { Session } from "../models/session.js";
import { ctrlWrapper } from "../helpers/index.js";
import { REFRESH_DURATION_SEC } from "../constants/index.js";

const cookieOptions = {
  httpOnly: true,
  // sameSite: "none",
  // secure: true, //для продакшн
  maxAge: REFRESH_DURATION_SEC * 1000,
};

const register = async (req, res) => {
  await registerUser(req.body);

  res.status(201).json({
    status: "success",
    code: 201,
    message:
      "Користувач успішно зареєстрований. Перевірте ваш email для підтвердження.",
  });
};

const verifyEmail = async (req, res) => {
  const { verificationCode } = req.body;

  await verifyUserEmail(verificationCode);

  res.status(204).json({
    status: "success",
    code: 204,
    message: "Email успішно підтверджено",
  });
};

const resendVerificationEmail = async (req, res) => {
  const { email } = req.body;
  await resendVerificationUserEmail(email);

  res.status(200).json({
    status: "success",
    code: 200,
    message: "Лист для підтвердження реєстрації успішно відправлено",
  });
};

const login = async (req, res) => {
  const result = await loginUser(req.body);

  res.cookie("refreshToken", result.session.refreshToken, cookieOptions);

  res.status(200).json({
    status: "success",
    code: 200,
    message: "Успішний вхід",
    data: {
      user: result.user,
      token: result.session.accessToken,
    },
  });
};

const logout = async (req, res) => {
  await Session.findByIdAndDelete(req.session?._id);
  res.clearCookie("refreshToken", cookieOptions);
  res.status(204).send();
};

const current = async (req, res) => {
  console.log("req.baseurl", req.baseUrl);
  console.log("req.path", req.path);
  console.log("req", req);

  res.json({
    status: "success",
    code: 200,
    message: "Поточний користувач отриманий",
    data: {
      user: { ...req.user, totalBalance: req.user.totalBalance / 100 || 0 },
      token: req.session?.accessToken,
    },
  });
};

const refresh = async (req, res) => {
  const { refreshToken: oldRefreshToken } = req.cookies;

  if (!oldRefreshToken) {
    return res.status(401).json({
      status: "error",
      code: 401,
      message: "Відсутній refresh token або session ID",
    });
  }

  const result = await refreshSession(oldRefreshToken);

  res.cookie("refreshToken", result.session.refreshToken, cookieOptions);

  res.status(200).json({
    status: "success",
    code: 200,
    message: "Сесія оновлена",
    data: {
      user: result.user,
      token: result.session.accessToken,
    },
  });
};

export default {
  register: ctrlWrapper(register),
  verifyEmail: ctrlWrapper(verifyEmail),
  resendVerificationEmail: ctrlWrapper(resendVerificationEmail),
  login: ctrlWrapper(login),
  logout: ctrlWrapper(logout),
  current: ctrlWrapper(current),
  refresh: ctrlWrapper(refresh),
};